Add compute capability for SCC V100s

+++ src/FloatCUDAInOneWeekend/material.h
@@ -19,7 +19,7 @@ struct material {
     MaterialType type;
     color albedo;
     float fuzz;
-    double refraction_index;
+    float refraction_index;
 
     __host__ __device__ material() {}
     // lambertian constructor
